+++ client/task_runner_test.go
@@ -31,7 +31,10 @@ func testLogger() *log.Logger {
 }
 
 func prefixedTestLogger(prefix string) *log.Logger {
-	return log.New(os.Stderr, prefix, log.LstdFlags)
+	if testing.Verbose() {
+		return log.New(os.Stderr, prefix, log.LstdFlags)
+	}
+	return log.New(ioutil.Discard, "", 0)
 }
 
 type MockTaskStateUpdater struct {